requirements.txt
aniso8601==0.92
Flask==0.10.1
Flask-Cors==1.8.0
flask-marshmallow==0.4.0
Flask-RESTful==0.2.12
Flask-Script==2.0.5
flask-mongoengine==0.9.5
mongoengine==0.20.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
marshmallow==1.2.2
mock==1.0.1
python-dotenv==0.19.2
pytz==2014.10
requests==2.3.0
six==1.9.0
Werkzeug==0.10.1
